+++ workspaces/scorecard/plugins/scorecard-backend-module-github/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @red-hat-developer-hub/backstage-plugin-scorecard-backend-module-github
 
+## 2.8.1
+
+### Patch Changes
+
+- 7ead71c: Correct default threshold documentation and add missing threshold documentation in scorecard backend module READMEs to match provider code defaults.
+  - @red-hat-developer-hub/backstage-plugin-scorecard-common@2.8.1
+  - @red-hat-developer-hub/backstage-plugin-scorecard-node@2.8.1
+
+## 2.8.0
+
+### Minor Changes
+
+- 8c85bd4: Backstage version bump to v1.51.1
+
+### Patch Changes
+
+- Updated dependencies [efb4c4f]
+- Updated dependencies [8c85bd4]
+  - @red-hat-developer-hub/backstage-plugin-scorecard-node@2.8.0
+  - @red-hat-developer-hub/backstage-plugin-scorecard-common@2.8.0
+
 ## 2.7.9
 
 ### Patch Changes